refactor: move spans interval logic into interval.go and always send a valid interval
